CyberSecurity/inspector-image/requirements.txt
Pillow==9.2.0
ExifRead==3.0.0
